requirements.txt
pdfplumber
PyMuPDF
fpdf
reportlab
PyPDF2
openai
google-cloud-vision
pytesseract
pdfminer.six
paddleocr
google-generativeai>=0.3.2
python-dotenv
streamlit